Event subscribers are invoked when performing Elasticsearch operations.
When Elasticsearch index plugin receives an index-able object (entity or array) for indexing,
Elasticsearch operation event is emited to allow event subscribers to decide if object needs
to be indexed.

Before request is sent to Elasticserach, Elasticsearch operation request event is emited to
allow event subscribers to modify the request execution callback and/or its parameters.